Isolating upload file process into a separated hook
